requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.11.1
amqp==5.1.1
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
beautifulsoup4==4.12.3
billiard==3.6.4.0
black==22.10.0
blinker==1.6.2
celery==5.2.7
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
commonmark==0.9.1
coverage==7.2.7
dataclasses-json==0.5.13
decorator==5.1.1
Deprecated==1.2.14
dnspython
duckduckgo-search==3.9.3
email-validator==2.0.0.post2
Faker==15.1.1
flake8==5.0.4
Flask==2.2.2
Flask-Babel==2.0.0
Flask-DB==0.3.2
Flask-DebugToolbar==0.13.1
Flask-HTTPAuth==4.8.0
Flask-Limiter==2.6.3
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Secrets==0.1.0
Flask-SQLAlchemy==3.0.3
Flask-Static-Digest==0.2.1
Flask-WTF==1.0.1
frozenlist==1.4.0
fsspec==2023.6.0
google-search-results==2.4.2
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore
idna==3.4
importlib-resources==6.0.0
infinity==1.5
iniconfig==2.0.0
intervals==0.9.2
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.1
limits==3.5.0
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.1
openai==1.37.2
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.1
pathspec==0.11.1
platformdirs==3.9.1
pluggy==1.2.0
prompt-toolkit==3.0.39
psycopg==3.1.12
psycopg-binary==3.1.12
psycopg2-binary==2.9.6
py==1.11.0
pycodestyle==2.9.1
pycountry==24.6.1
pydantic==1.10.12
pyflakes==2.5.0
Pygments==2.15.1
pytest==7.1.3
pytest-cov==4.0.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
redis==4.3.4
regex==2023.6.3
requests==2.29.0
rich==12.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
SQLAlchemy-Utils==0.41.1
stripe==9.7.0
tabulate==0.9.0
tenacity==8.2.2
tiktoken==0.4.0
tomli==2.0.1
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
validators==0.20.0
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.2.2
wrapt==1.15.0
WTForms==3.0.1
WTForms-Alchemy==0.18.0
WTForms-Components==0.10.5
yarl==1.9.2
